Adding polltime as configuration in entity manager

+++ include/IpmbSensor.hpp
@@ -230,7 +230,7 @@ struct IpmbSensor : public Sensor
     std::vector<uint8_t> initData;
 
     std::string versionTypeName;
-    uint8_t pollTimeValue;
+    int pollTimeValue;
     uint8_t index;
 
     ReadingFormat readingFormat;
